Refactor: Improve retry logic and UI error handling
This commit introduces several enhancements to the retry mechanism and UI feedback for network operations:

- **Refined Retry Logic in `NetworkBoundResourceFlow`:**
    - Replaced `retryWhen` with an internal `while` loop for more direct control over retry attempts and delay calculations.
    - Now emits `Resource.Loading` before each retry attempt to provide better UI feedback.
    - `Resource.Error` now includes an `isRetrying` flag, indicating to the UI whether another attempt will be made.
- **Enhanced `UiState`:**
    - Added an `isRetrying` boolean field to `UiState` to signal to Composables if a retry is in progress after an error.
- **Improved UI Error Display:**
    - `AnimeEpisodesList` and `CharacterList` now check the `isRetrying` flag in `UiState.errorMessage`.
    - If `isRetrying` is true, a "Retrying shortly..." message and a small `CircularProgressIndicator` are displayed.
    - If `isRetrying` is false (final error), a "Try Again" button is shown (currently a placeholder).
    - `AnimeEpisodesList` now displays a more informative "No episodes found" message with an icon when the episodes list is empty, rather than just showing nothing.
- **ViewModel Updates:**
    - `AnimeDetailViewModel` now correctly passes the `isRetrying` flag from `Resource.Error` to `UiState.error`.
- **Type Correction in `RetryUtils`:**
    - Changed the `attempt` parameter type in `calculateBackoffDelay` from `Long` to `Int` for consistency with retry attempt counts.

+++ core/common/src/main/java/com/nightfire/tonkotsu/core/common/RetryUtils.kt
@@ -18,7 +18,7 @@ fun isRateLimitOrNetworkError(cause: Throwable): Boolean {
     }
 }
 
-fun calculateBackoffDelay(attempt: Long): Long {
+fun calculateBackoffDelay(attempt: Int): Long {
     val exponent = (attempt - 1).toInt()
     return (RetryConfig.INITIAL_BACKOFF_MILLIS * RetryConfig.BACKOFF_FACTOR.toDouble().pow(exponent)).toLong()
 }
